+++ src/lib/util/my_list.ml
@@ -74,3 +74,11 @@ let rec is_sorted cmp l =
   match l with
   | x :: y :: xs -> cmp x y <= 0 && is_sorted cmp (y :: xs)
   | [_] | [] -> true
+
+let rec iter_pair f (l1, l2) =
+  match l1, l2 with
+  | [], [] -> ()
+  | hd1 :: tl1, hd2 :: tl2 ->
+    f (hd1, hd2);
+    iter_pair f (tl1, tl2)
+  | _ -> invalid_arg "iter_pair"